requirements.txt
streamlit>=1.32.0
stmol==0.0.9
py3Dmol>=2.0.0
biotite>=0.39.0
ipywidgets>=8.0.0
ipython_genutils>=0.2.0
requests>=2.31.0
altair>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0